requirements.txt
python~=3.9.7
numpy~=1.21.2
pandas~=1.3.5
timm==0.5.4
tqdm~=4.62.3
fastai~=2.5.3
pillow~=9.0.0
sklearn~=1.0.2
matplotlib~=3.5.0
cudatoolkit~=11.3.1
torchvision~=0.11.2
